Section4-ContinuousTraining-Airflow-Composer/assignment/prediction-serving-app-cloudrun/requirements.txt
Flask>=3.0.0
gunicorn>=22.0.0
scikit-learn>=1.5.0
pandas>=2.2.0
pyarrow>=15.0.0
joblib>=1.3.0
google-cloud-storage>=2.14.0
google-cloud-logging>=3.9.0
